+++ packages/sanity/src/core/studio/copyPaste/transferValue.ts
